Show variables of anonymous objects as code-minings in the debugger
`object <type> {}` and `map <mapping>` create variables that are named
"$temp_<n>" and "temp<n>". Before this change there was no way to know
what object each variable refers to. With this change the variable name
is shown in the editor as a code-mining, looking like this for example:
`object $temp_1 RDB::TableColumn {}`
`map temp3.persistentClass2table()`