Hide internal variables in the variables view
This commit hides the `__qvtresult__`, `__result__` and `_intermediate`
variables in the variables view.

`__qvtresult__` is the temp variable used for accumulating the results
of a `map` call.

`__result__` is the temp variable used for accumulating the results for
loop expressions like `exists()`.

`_intermediate` is an internal variable that holds intermediate classes.
 As intermediate classes are shown as the type for variables like
 `result` there is no need to show this internal variable.